requirements/base.txt
# Django
Django==2.1.7

# Django Rest Framework
djangorestframework==3.9.1
djoser==1.4.0


# Django Cors Headers
django-cors-headers==2.4.1

# Ck Editor
django-ckeditor==5.7.1
# django-js-asset==1.2.2
# pkg-resources==0.0.0


# Crispy Form
django-crispy-forms==1.7.2
# django-templated-mail==1.1.1

# Django Filter
django-filter==2.1.0

# reCAPTCHA
django-recaptcha==2.0.4
